requirements.txt
langchain>=0.2.12
langchain-openai>=0.1.7
openai>=1.47.0
pydantic>=2.7.0
python-dotenv>=1.0.1
